+++ autonomi/src/native/client/transfers.rs
@@ -1,10 +1,29 @@
-use crate::wallet::MemWallet;
-use crate::Client;
+use crate::client::ClientWrapper;
+use crate::native::client::NativeClient;
+use crate::native::wallet;
+use crate::native::wallet::MemWallet;
+use crate::VERIFY_STORE;
+use libp2p::{
+    futures::future::join_all,
+    kad::{Quorum, Record},
+    PeerId,
+};
+use sn_networking::{
+    GetRecordCfg, GetRecordError, Network, NetworkError, PutRecordCfg, VerificationKind,
+};
+use sn_protocol::{
+    storage::{try_serialize_record, RecordKind, RetryStrategy, SpendAddress},
+    NetworkAddress, PrettyPrintRecordKey,
+};
+use sn_transfers::CashNote;
+use sn_transfers::Payment;
+use sn_transfers::UniquePubkey;
+use sn_transfers::{HotWallet, SignedSpend};
 use sn_transfers::{MainPubkey, NanoTokens};
 use sn_transfers::{SpendReason, Transfer};
-
-use sn_transfers::UniquePubkey;
 use std::collections::BTreeSet;
+use std::collections::HashSet;
+use xor_name::XorName;
 
 #[derive(Debug, thiserror::Error)]
 pub enum SendSpendsError {
@@ -23,7 +42,7 @@ pub enum TransferError {
     #[error("Wallet error: {0:?}")]
     WalletError(#[from] wallet::error::WalletError),
     #[error("Network error: {0:?}")]
-    NetworkError(#[from] sn_networking::NetworkError),
+    NetworkError(#[from] NetworkError),
 }
 
 #[derive(Debug, thiserror::Error)]
@@ -33,29 +52,6 @@ pub enum CashNoteError {
     #[error("Failed to get spend: {0:?}")]
     FailedToGetSpend(String),
 }
-
-use libp2p::{
-    futures::future::join_all,
-    kad::{Quorum, Record},
-    PeerId,
-};
-use sn_networking::{
-    GetRecordCfg, GetRecordError, Network, NetworkError, PutRecordCfg, VerificationKind,
-};
-use sn_protocol::{
-    storage::{try_serialize_record, RecordKind, RetryStrategy, SpendAddress},
-    NetworkAddress, PrettyPrintRecordKey,
-};
-use sn_transfers::Payment;
-use sn_transfers::{HotWallet, SignedSpend};
-use xor_name::XorName;
-use crate::client::ClientWrapper;
-use crate::native::client::NativeClient;
-use crate::native::wallet::MemWallet;
-use crate::native::{wallet, Client};
-use crate::VERIFY_STORE;
-use sn_transfers::CashNote;
-use std::collections::HashSet;
 
 #[derive(Debug, thiserror::Error)]
 pub enum SendError {
