+++ autonomi/src/native/mod.rs
@@ -1,6 +1,5 @@
-pub use client::Client;
+pub use crate::client::Client;
 
 pub mod client;
-mod secrets;
 #[cfg(feature = "transfers")]
 mod wallet;
